move function-constant-pinning to experimental
