+++ xdsl/tools/command_line_tool.py
@@ -526,7 +526,7 @@ def get_all_passes() -> dict[str, Callable[[], type[ModulePass]]]:
         return convert_scf_to_riscv_scf.ConvertScfToRiscvPass
 
     def get_function_constant_pinning():
-        from xdsl.transforms.function_constant_pinning import (
+        from xdsl.transforms.experimental.function_constant_pinning import (
             FunctionConstantPinningPass,
         )
